Move urlParamsInEditMode to ChangeController
