+++ libraries/classes/Controllers/Table/ChangeController.php
@@ -24,6 +24,7 @@ use function is_string;
 use function str_contains;
 use function strlen;
 use function strpos;
+use function trim;
 
 /**
  * Displays form for editing and inserting new table rows.
@@ -186,10 +187,7 @@ class ChangeController extends AbstractController
 
         $GLOBALS['urlParams']['db'] = $GLOBALS['db'];
         $GLOBALS['urlParams']['table'] = $GLOBALS['table'];
-        $GLOBALS['urlParams'] = $this->insertEdit->urlParamsInEditMode(
-            $GLOBALS['urlParams'],
-            $GLOBALS['where_clause_array'],
-        );
+        $GLOBALS['urlParams'] = $this->urlParamsInEditMode($GLOBALS['urlParams'], $GLOBALS['where_clause_array']);
 
         $GLOBALS['has_blob_field'] = false;
         foreach ($GLOBALS['table_columns'] as $column) {
@@ -307,5 +305,28 @@ class ChangeController extends AbstractController
         }
 
         $this->response->addHTML($htmlOutput);
+    }
+
+    /**
+     * Add some url parameters
+     *
+     * @param mixed[] $urlParams        containing $db and $table as url parameters
+     * @param mixed[] $whereClauseArray where clauses array
+     *
+     * @return mixed[] Add some url parameters to $url_params array and return it
+     */
+    public function urlParamsInEditMode(
+        array $urlParams,
+        array $whereClauseArray,
+    ): array {
+        foreach ($whereClauseArray as $whereClause) {
+            $urlParams['where_clause'] = trim($whereClause);
+        }
+
+        if (! empty($_POST['sql_query'])) {
+            $urlParams['sql_query'] = $_POST['sql_query'];
+        }
+
+        return $urlParams;
     }
 }
